Move mapping tests into own dir; start refactoring
