+++ test/unit/data/model/mapping/test_install_model_mapping.py
